// src/app/modules/master/shift/constant/workingdays.enum.ts
/* eslint-disable no-magic-numbers */
// eslint-disable-next-line no-shadow
export enum WorkingDaysValues {
	MonValue = 1,
	TueValue = 2,
	WedValue = 3,
	ThuValue = 4,
	FriValue = 5,
	SatValue = 6,
	SunValue = 7
  };
export enum WorkingDays {
	  Mon = 'Mon',
	  Tue = 'Tue',
	  Wed = 'Wed',
	  Thu = 'Thu',
	  Fri = 'Fri',
	  Sat = 'Sat',
	  Sun = 'Sun'
  };
export enum PunchReportDaySetting {
	  PunchIn = 'Punch In',
	  PunchOut = 'Punch Out',
	  In = 'In',
	  Out = 'Out'
  }